Expose protocol::buf::gap module
This is useful as one of the primary pain points of working with the
Kafka API is needing to manually encode the length delimiters. Often
times it is nice to first reserve the necessary slot, continue to encode
header & response body, then go back and fill the slot.

This pattern is used quite a lot internal to this crate. It is a useful
pattern outside of this crate as well.